+++ src/index.ts
@@ -3,4 +3,5 @@ export { IntentProvider };
 export * from './types/provider';
 export * from './types/jsonrpc';
 export * from './types/recentTxs';
-export * from './types/tokenPrice'; 
+export * from './types/tokenPrice';
+export * from './types/tokenBalances';